requirements.txt
boto3==1.37.9
botocore==1.37.9
jmespath==1.0.1
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
